apps/ice/plugins/oxml/oxml_docProps_test.py
#!/usr/bin/env python
#
#    Copyright (C) 2010  Distance and e-Learning Centre,
#    University of Southern Queensland
#
#    This program is free software; you can redistribute it and/or modify
#    it under the terms of the GNU General Public License as published by
#    the Free Software Foundation; either version 2 of the License, or
#    (at your option) any later version.
#
#    This program is distributed in the hope that it will be useful,
#    but WITHOUT ANY WARRANTY; without even the implied warranty of
#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#    GNU General Public License for more details.
#
#    You should have received a copy of the GNU General Public License
#    along with this program; if not, write to the Free Software
#    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
#

from unittest import TestCase
import sys

from oxml_docProps import DocProps

# docProps/core.xml
testFile = "testData/core.xml"


class DocPropsTest(TestCase):

    def setUp(self):
        f = open(testFile, "rb")
        self.coreXmlStr = f.read()
        f.close()

    def tearDown(self):
        pass

    def testTitle(self):
        docProps = DocProps(self.coreXmlStr)
        #print docProps.title
        self.assertEquals(docProps.title, "TestImages")

    def testGetProperty(self):
        docProps = DocProps(self.coreXmlStr)
        self.assertEquals(docProps.getProperty("dc:title"), "TestImages")
        self.assertEquals(docProps.getProperty("dc:creator"), "Test User")
        self.assertEquals(docProps.getProperty("cp:lastModifiedBy"), "Test User")
        self.assertEquals(docProps.getProperty("dcterms:created"), u"2010-04-12T03:13:00Z")
        self.assertEquals(docProps.getProperty("dcterms:modified"), u"2010-04-22T04:03:00Z")




def runUnitTests(locals):
    print "\n\n\n\n"
    if sys.platform=="cli":
        import clr
        import System.Console
        System.Console.Clear()
        print "---- Testing under IronPython ----"
    else:
        print "---- Testing ----"

    # Run only the selected tests
    args = list(sys.argv)
    sys.argv = sys.argv[:1]
    args.pop(0)
    runTests = args
    runTests = [ i.lower().strip(", ") for i in runTests]
    runTests = ["test"+i for i in runTests if not i.startswith("test")] + \
                [i for i in runTests if i.startswith("test")]
    if runTests!=[]:
        testClasses = [i for i in locals.values() \
                        if hasattr(i, "__bases__") and \
                            (TestCase in i.__bases__)]
        testing = []
        for x in testClasses:
            l = dir(x)
            l = [ i for i in l if i.startswith("test") and callable(getattr(x, i))]
            for i in l:
                if i.lower() not in runTests:
                    delattr(x, i)
                else:
                    testing.append(i)
        x = None
        num = len(testing)
        if num<1:
            print "No selected tests found! - %s" % str(args)[1:-1]
        elif num==1:
            print "Running selected test - %s" % (str(testing)[1:-1])
        else:
            print "Running %s selected tests - %s" % (num, str(testing)[1:-1])
    from unittest import main
    main()


if __name__=="__main__":
    runUnitTests(locals())
    sys.exit(0)



